Break up integration tests
