+++ tests/open_alchemy/test_integration/test_database.py
@@ -1,81 +1,12 @@
-"""Integration tests."""
+"""Integration tests against database."""
 
-import json
 import typing
-from unittest import mock
 
 import pytest
 import sqlalchemy
-import yaml
 from sqlalchemy.ext import declarative
 
 import open_alchemy
-
-
-@pytest.mark.integration
-def test_to_from_dict():
-    """
-    GIVEN specification that has schema with a single property
-    WHEN model is defined based on schema and constructed using from_dict
-    THEN when to_dict is called the construction dictionary is returned.
-    """
-    model_factory = open_alchemy.init_model_factory(
-        base=mock.MagicMock,
-        spec={
-            "components": {
-                "schemas": {
-                    "Table": {
-                        "properties": {"column": {"type": "integer"}},
-                        "x-tablename": "table",
-                        "type": "object",
-                    }
-                }
-            }
-        },
-    )
-    model = model_factory(name="Table")
-
-    # COnstructing and turning back to dictionary
-    model_dict = {"column": 1}
-    instance = model.from_dict(**model_dict)
-    assert instance.to_dict() == model_dict
-
-
-@pytest.mark.integration
-def test_to_from_dict_relationship():
-    """
-    GIVEN specification that has a schema with a relationship
-    WHEN model is defined based on schema and constructed using from_dict
-    THEN when to_dict is called the construction dictionary is returned.
-    """
-    model_factory = open_alchemy.init_model_factory(
-        base=mock.MagicMock,
-        spec={
-            "components": {
-                "schemas": {
-                    "RefTable": {
-                        "properties": {"id": {"type": "integer"}},
-                        "x-tablename": "table",
-                        "type": "object",
-                    },
-                    "Table": {
-                        "properties": {
-                            "column": {"$ref": "#/components/schemas/RefTable"}
-                        },
-                        "x-tablename": "table",
-                        "type": "object",
-                    },
-                }
-            }
-        },
-    )
-    model_factory(name="RefTable")
-    model = model_factory(name="Table")
-
-    # COnstructing and turning back to dictionary
-    model_dict = {"column": {"id": 1}}
-    instance = model.from_dict(**model_dict)
-    assert instance.to_dict() == model_dict
 
 
 @pytest.mark.parametrize(
@@ -581,281 +512,3 @@ def test_database_ref_all_of(engine, sessionmaker, spec):
     # Querying session
     queried_model = session.query(model).first()
     assert queried_model.column == 1
-
-
-@pytest.mark.parametrize(
-    "schema_additions, sql, expected_contents",
-    [
-        (
-            {"x-composite-unique": ["id"]},
-            "SELECT sql FROM sqlite_master WHERE name='table'",
-            ["UNIQUE (id)"],
-        ),
-        (
-            {"x-composite-unique": [["id"], ["column"]]},
-            "SELECT sql FROM sqlite_master WHERE name='table'",
-            ["UNIQUE (id)", 'UNIQUE ("column")'],
-        ),
-        (
-            {"x-composite-unique": {"columns": ["id"]}},
-            "SELECT sql FROM sqlite_master WHERE name='table'",
-            ["UNIQUE (id)"],
-        ),
-        (
-            {"x-composite-unique": {"name": "id", "columns": ["id"]}},
-            "SELECT sql FROM sqlite_master WHERE name='table'",
-            ["CONSTRAINT id UNIQUE (id)"],
-        ),
-        (
-            {"x-composite-unique": [{"columns": ["id"]}, {"columns": ["column"]}]},
-            "SELECT sql FROM sqlite_master WHERE name='table'",
-            ["UNIQUE (id)", 'UNIQUE ("column")'],
-        ),
-        (
-            {"x-composite-index": ["id"]},
-            "SELECT sql FROM sqlite_master WHERE type='index'",
-            ['INDEX ix_table_id ON "table" (id)'],
-        ),
-        (
-            {"x-composite-index": [["id"], ["column"]]},
-            "SELECT sql FROM sqlite_master WHERE type='index'",
-            [
-                'INDEX ix_table_id ON "table" (id)',
-                'INDEX ix_table_column ON "table" ("column")',
-            ],
-        ),
-        (
-            {"x-composite-index": {"expressions": ["id"]}},
-            "SELECT sql FROM sqlite_master WHERE type='index'",
-            ['INDEX ix_table_id ON "table" (id)'],
-        ),
-        (
-            {"x-composite-index": {"name": "id", "expressions": ["id"]}},
-            "SELECT sql FROM sqlite_master WHERE type='index'",
-            ['INDEX id ON "table" (id)'],
-        ),
-        (
-            {"x-composite-index": {"expressions": ["id"], "unique": True}},
-            "SELECT sql FROM sqlite_master WHERE type='index'",
-            ['UNIQUE INDEX ix_table_id ON "table" (id)'],
-        ),
-        (
-            {
-                "x-composite-index": [
-                    {"expressions": ["id"]},
-                    {"expressions": ["column"]},
-                ]
-            },
-            "SELECT sql FROM sqlite_master WHERE type='index'",
-            [
-                'INDEX ix_table_id ON "table" (id)',
-                'INDEX ix_table_column ON "table" ("column")',
-            ],
-        ),
-    ],
-    ids=[
-        "unique array",
-        "unique multiple array",
-        "unique object",
-        "unique object name",
-        "unique multiple object",
-        "index array single",
-        "index multiple array",
-        "index object",
-        "index object name",
-        "index object unique",
-        "index multiple object",
-    ],
-)
-@pytest.mark.integration
-def test_table_args_unique(engine, schema_additions, sql, expected_contents):
-    """
-    GIVEN schema, additional properties for schema, sql to execute and expected
-        contents
-    WHEN models are constructed
-    THEN when sql is executed the expected contents are in the result.
-    """
-    # Defining schema
-    spec = {
-        "components": {
-            "schemas": {
-                "Table": {
-                    "properties": {
-                        "id": {"type": "integer", "x-primary-key": True},
-                        "column": {"type": "integer"},
-                    },
-                    "x-tablename": "table",
-                    "type": "object",
-                    **schema_additions,
-                }
-            }
-        }
-    }
-    # Creating model factory
-    base = declarative.declarative_base()
-    model_factory = open_alchemy.init_model_factory(spec=spec, base=base)
-    model_factory(name="Table")
-
-    # Creating models
-    base.metadata.create_all(engine)
-
-    # Query schema
-    results_list = list(str(result) for result in engine.execute(sql))
-    results = "\n".join(results_list)
-    for expected_content in expected_contents:
-        assert expected_content in results
-
-
-BASIC_SPEC = {
-    "components": {
-        "schemas": {
-            "Table": {
-                "properties": {"column": {"type": "integer", "x-primary-key": True}},
-                "x-tablename": "table",
-                "type": "object",
-            }
-        }
-    }
-}
-
-
-@pytest.mark.integration
-def test_init_json(engine, sessionmaker, tmp_path):
-    """
-    GIVEN specification stored in a JSON file
-    WHEN init_json is called with the file
-    THEN a valid model factory is returned.
-    """
-    # Generate spec file
-    directory = tmp_path / "specs"
-    directory.mkdir()
-    spec_file = directory / "spec.json"
-    spec_file.write_text(json.dumps(BASIC_SPEC))
-
-    # Creating model factory
-    base, model_factory = open_alchemy.init_json(str(spec_file), define_all=False)
-    model = model_factory(name="Table")
-
-    # Creating models
-    base.metadata.create_all(engine)
-    # Creating model instance
-    value = 0
-    model_instance = model(column=value)
-    session = sessionmaker()
-    session.add(model_instance)
-    session.flush()
-
-    # Querying session
-    queried_model = session.query(model).first()
-    assert queried_model.column == value
-
-
-@pytest.mark.integration
-def test_init_yaml(engine, sessionmaker, tmp_path):
-    """
-    GIVEN specification stored in a YAML file
-    WHEN init_yaml is called with the file
-    THEN a valid model factory is returned.
-    """
-    # Generate spec file
-    directory = tmp_path / "specs"
-    directory.mkdir()
-    spec_file = directory / "spec.yaml"
-    spec_file.write_text(yaml.dump(BASIC_SPEC))
-
-    # Creating model factory
-    base, model_factory = open_alchemy.init_yaml(str(spec_file), define_all=False)
-    model = model_factory(name="Table")
-
-    # Creating models
-    base.metadata.create_all(engine)
-    # Creating model instance
-    value = 0
-    model_instance = model(column=value)
-    session = sessionmaker()
-    session.add(model_instance)
-    session.flush()
-
-    # Querying session
-    queried_model = session.query(model).first()
-    assert queried_model.column == value
-
-
-@pytest.mark.integration
-def test_init_yaml_import_error():
-    """
-    GIVEN yaml package is not available
-    WHEN init_yaml is called
-    THEN ImportError is raised.
-    """
-    with mock.patch.dict("sys.modules", {"yaml": None}):
-        with pytest.raises(ImportError):
-            open_alchemy.init_yaml("some file")
-
-
-@pytest.mark.integration
-def test_import_base_initial():
-    """
-    GIVEN
-    WHEN
-    THEN ImportError is raised when Base is imported from open_alchemy.models.
-    """
-    # pylint: disable=import-error,import-outside-toplevel,unused-import
-    with pytest.raises(ImportError):
-        from open_alchemy.models import Base  # noqa: F401
-
-
-@pytest.mark.integration
-def test_import_base(tmp_path):
-    """
-    GIVEN specification file
-    WHEN init_yaml is called with the specification file
-    THEN Base can be imported from open_alchemy.models.
-    """
-    # pylint: disable=import-error,import-outside-toplevel,unused-import
-    # Generate spec file
-    directory = tmp_path / "specs"
-    directory.mkdir()
-    spec_file = directory / "spec.yaml"
-    spec_file.write_text(yaml.dump(BASIC_SPEC))
-
-    # Creating model factory
-    open_alchemy.init_yaml(str(spec_file))
-
-    from open_alchemy.models import Base  # noqa: F401
-
-
-@pytest.mark.integration
-def test_import_model(engine, sessionmaker, tmp_path):
-    """
-    GIVEN specification stored in a YAML file
-    WHEN init_yaml is called with the file
-    THEN the model is importable from open_alchemy.models.
-    """
-    # pylint: disable=import-error,import-outside-toplevel
-    # Generate spec file
-    directory = tmp_path / "specs"
-    directory.mkdir()
-    spec_file = directory / "spec.yaml"
-    spec_file.write_text(yaml.dump(BASIC_SPEC))
-
-    # Creating model factory
-    open_alchemy.init_yaml(str(spec_file), define_all=True)
-
-    # Creating models
-    from open_alchemy.models import Base
-
-    Base.metadata.create_all(engine)
-
-    # Creating model instance
-    from open_alchemy.models import Table
-
-    value = 0
-    model_instance = Table(column=value)
-    session = sessionmaker()
-    session.add(model_instance)
-    session.flush()
-
-    # Querying session
-    queried_model = session.query(Table).first()
-    assert queried_model.column == value
